refactor: вынести функционал профайла в модули
