+++ .eslintrc.json
@@ -39,7 +39,6 @@
     ],
     "react/react-in-jsx-scope": "off",
     "react/jsx-props-no-spreading": "off",
-    "react/jsx-no-bind": "off",
     "no-use-before-define": [
       "error",
       {
